refactor: extract `transformComponent` for standalone usage
